requirements.txt
pytest==6.1.2
selenium==3.141.0
requests==2.24.0
pytest-xdist==2.1.0
allure-pytest==2.8.19
mysql-connector-python==8.0.22
